refactor + cleanup
