+++ packages/ssx-sdk/src/modules/Storage/KeplerDataVault.ts
@@ -1,0 +1,73 @@
+import { IEncryption } from '../Encryption';
+import { IUserAuthorization } from '../../';
+import { IDataVault } from './interfaces';
+import KeplerStorage from './KeplerStorage';
+
+export class KeplerDataVault extends KeplerStorage implements IDataVault {
+  private encryption: IEncryption;
+
+  /**
+   * Constructs a new instance of KeplerDataVault.
+   * @param config - An object containing configuration properties for Kepler.
+   * @param userAuth - An instance of IUserAuthorization for user authorization.
+   * @param encryption - An instance of IEncryption for encryption/decryption ops.
+   */
+  constructor(
+    config: any,
+    userAuth: IUserAuthorization,
+    encryption: IEncryption
+  ) {
+    super(config, userAuth);
+    this.encryption = encryption;
+  }
+
+  /**
+   * Retrieves the encrypted stored value for the specified key,
+   * decrypts it, and returns the decrypted value.
+   * @param key - The unique identifier for the stored value.
+   * @returns A Promise that resolves to the decrypted value associated
+   * with the given key or null if the key does not exist.
+   */
+  public async get(key: string): Promise<any> {
+    const { data } = await super.get(key);
+    // TODO: check metadata for encryption type
+    if (data) {
+      const decrypted = await this.encryption.decrypt(data);
+      return { data: decrypted };
+    }
+    return { data };
+  }
+
+  /**
+   * Stores a value with the specified key, encrypting it before storage.
+   * @param key - The unique identifier for the stored value.
+   * @param value - The value to store under the given key with encryption.
+   * @returns A Promise that resolves when the operation is complete.
+   */
+  public async put(key: string, value: any): Promise<void> {
+    // TODO: put metadata for encryption type
+    const encryptedData = await this.encryption.encrypt(value);
+    return super.put(key, encryptedData);
+  }
+
+  /**
+   * Retrieves the unencrypted stored value associated with the specified key.
+   * @param key - The unique identifier for the stored value.
+   * @returns A Promise that resolves to the unencrypted value associated with the given key or undefined if the key does not exist.
+   */
+  public unencrypted_get(key: string): Promise<any> {
+    return super.get(key);
+  }
+
+  /**
+   * Stores a value without encryption with the specified key.
+   * @param key - The unique identifier for the stored value.
+   * @param value - The value to store under the given key without encryption.
+   * @returns A Promise that resolves when the operation is complete.
+   */
+  public unencrypted_put(key: string, value: any): Promise<void> {
+    return super.put(key, value);
+  }
+}
+
+export default KeplerDataVault;
