--- packages/ssx-sdk/src/modules/common.ts
@@ -1,12 +1,0 @@
-export interface ResourceError {
-  type: string;
-  message: string;
-  data?: any;
-}
-
-export interface ResourceResult {
-  success: boolean;
-  data?: any;
-  error?: ResourceError;
-  message?: string;
-}
